pyproject.toml
[tool.poetry]
name = "hetznerbot"
version = "1.0.0"
description = "A bot for finding the best offers from hetzner"
authors = ["Arne Beer <contact@arne.beer>"]
license = "MIT"
repository = "https://github.com/nukesor/hetznerbot"

[tool.poetry.dependencies]
python = ">=3.6"
python-telegram-bot = "^13"
sentry-sdk = "^1"
alembic = "^1"
SQLAlchemy = "^1"
sqlalchemy-utils = "^0.37"
psycopg2-binary = "^2"
toml = "^0.10"
requests = "^2"
dateparser = "^1"

[tool.poetry.dev-dependencies]
pytest-mock = "^3"
pytest = "^6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
